+++ src/main.cpp
@@ -7,7 +7,7 @@
 
 int main(int argc, char* argv[]){
     const string config_file_path = argv[CONFIG_FILE_INDEX];
-    map<string, map<string, int>> config = read_config(config_file_path);
+    map<string, map<string, float>> config = read_config(config_file_path);
     PVZ game = PVZ(config);
     game.run();
     return 0;
